NTLWAY/back end/backend/requirements.txt
djangorestframework==3.12.2
beautifulsoup4==4.8.2
djangorestframework-simplejwt
django-cors-headers==3.5.0
django-cors-middleware==1.5.0

# for the devis creation
docxtpl==0.11.2
PyPDF2==1.26.0
comtypes==1.1.8 # to remove works only on windows
docx2pdf==0.1.7


# for the conversion

PyVirtualDisplay==1.3.2
pdfkit==0.6.1

#############
requests==2.26.0

drf_pdf
reportlab

WeasyPrint